requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.104.1
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
huggingface-hub==0.17.3
idna==3.10
joblib==1.5.1
keras==2.15.0
libclang==18.1.1
lxml==6.0.0
Markdown==3.8.2
MarkupSafe==3.0.2
ml-dtypes==0.2.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pandas==2.1.3
protobuf==4.25.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic_core==2.14.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.3.2
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.14.1
tqdm==4.67.1
transformers==4.35.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
Werkzeug==3.1.3
wrapt==1.14.2
